ai_services/pyproject.toml
[project]
name = "ai-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
]
